bootstrap으로 페이지 꾸미기

+++ src/main/java/com/springboot/leadingbooks/services/dto/response/BookCreateResponseDto.java
@@ -14,7 +14,7 @@ public class BookCreateResponseDto {
     private String bWriter;
     private String bPublish;
     private Category category;
-    private int bCount;
+    private Long bCount;
 
     public static BookCreateResponseDto of(Book book) {
         return BookCreateResponseDto.builder()
